+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-07-02] [Make server more typesafe](https://github.com/rubriclab/state/commit/43a77f607fd0f2942e21f9634ba3953267de0d75)
 - [2025-07-02] [Bring back client-side cookie-set](https://github.com/rubriclab/state/commit/74f79bd2b7dfa32e834428591361e613076187fb)
 - [2025-07-01] [run format](https://github.com/RubricLab/state/commit/2ccbd6d49c92c4d11da9c73b9e83561ff4e08258)
 - [2025-07-01] [consolidate biome format script](https://github.com/RubricLab/state/commit/a4f45d0cd36cf2b7df80803f9bc23bff8925c0b3)